requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
pytesseract==0.3.10
transformers==4.36.0
torch==2.1.0
requests==2.31.0
python-multipart==0.0.6
psutil==5.9.6